Try to fix build on fresh gcc's

+++ protocols/oscar/liboscar/tasks/chatnavservicetask.cpp
@@ -103,7 +103,7 @@ bool ChatNavServiceTask::take( Transfer* transfer )
 void ChatNavServiceTask::onGo()
 {
     FLAP f =  { 0x02, 0, 0x00 };
-    SNAC s = { 0x000D, m_type, 0x0000, client()->snacSequence() };
+    SNAC s = { 0x000D, static_cast<unsigned short>(m_type), 0x0000, client()->snacSequence() };
     Buffer* b = new Buffer();
 
     Transfer* t = createTransfer( f, s, b );
